♻️ Refactor toast error to use composable
